Model A/B Plus compatibility with major simplification
Simplified header connetor to ONLY be a Raspberry Pi header connectorizer -
that is, it no longer takes speaker or audio (or the now deprecated USB power
adapter).  Speaker and audio will be handled with simpler cables to the
breadboard.  So, now the header only has 2 connectors: a header to the RasPi,
and a header for breadboard connections.

Also, a total redo of the labeling to have a much more clear grid layout,
separatating GPIO pin numbers from alternate pin functions.